#Author : mounicraju@gmail.com
% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/self-destruction.R
\docType{package}
\name{self-destruction}
\alias{self-destruction}
\alias{self-destruction-package}
\alias{self-destruction}
\alias{batch_self-destruction}
\title{Let a process kill itself}
\usage{
self-destruction(interval)

batch_self-destruction(interval)
}
\arguments{
\item{interval}{number of seconds to sleep}
}
\description{
Let a process kill itself
}
